Move FigureMaker to a module; add conda install instructions and environment.
